get and set

+++ Data-Structures-Ex2/Bridge.cpp
@@ -1,6 +1,6 @@
 #include "Bridge.h"
 
-Bridge::Bridge(int height, Bridge* next)
+Bridge::Bridge(float height, Bridge* next)
 {
 	_height = height;
 	_next = next;
