+++ Data-Structures-Ex2/Bridge.h
@@ -3,11 +3,11 @@
 class Bridge
 {
 private:
-	int _height;
+	float _height;
 	Bridge* _next;
 
 public:
-	Bridge(int height = 0, Bridge* next = nullptr);
+	Bridge(float height = 0, Bridge* next = nullptr);
 	void setHeight(int height) { _height = height; }
 	int getHeight() { return _height; }
 	void setNext(Bridge* next) { _next = next; }
